+++ worms/__main__.py
@@ -1,0 +1,3 @@
+import pytest
+import sys
+pytest.main(sys.argv + ['--pyargs', 'worms'])